tearDown driver!=null

+++ src/test/java/rbauction/tests/BaseTest_RBAuction.java
@@ -99,11 +99,13 @@ public class BaseTest_RBAuction {
         } catch (Exception e) {
             log("Test Rail was not updated: "+e);
         }
-        log("DEBUG: tearDown deleteCookies");
-        driver.manage().deleteAllCookies();
-        log("DEBUG: tearDown driver.quit - start");
-        driver.quit();
-        log("DEBUG: tearDown  driver.quit -successful");
+        if (driver!=null) {
+            log("DEBUG: tearDown deleteCookies");
+            driver.manage().deleteAllCookies();
+            log("DEBUG: tearDown driver.quit - start");
+            driver.quit();
+            log("DEBUG: tearDown  driver.quit -successful");
+        }
     }
 
     public WebDriver getDriver() {
